% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pacific_oyster_18S_V4.R
\docType{data}
\name{pacific_oyster_18S_V4}
\alias{pacific_oyster_18S_V4}
\title{pacific_oyster_18S_V4:}
\format{
A data frame
}
\usage{
data(pacific_oyster_18S_V4)
}
\description{
18S rRNA gene sequence of pacific oyster at V4 region that amplified by primers "TAReuk454FWD1 and TAReukREV3" (Stoeck et al., 2010).
This serves as an example of input fasta file for cas9.gRNA.oligo1, cas9.gRNA.oligo2, cas12a.gRNA.oligo1 and cas12a.gRNA.oligo2.
}
\examples{
data(pacific_oyster_18S_V4)
}
\references{
Stoeck T, Bass D, Nebel M, Christen R, Jones MD, Breiner HW, Richards TA. (2010)
Multiple marker parallel tag environmental DNA sequencing reveals a highly complex eukaryotic community in marine anoxic water
Mol. Ecol. 19, 21–31.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/20331767}{PubMed})
}
\keyword{pacific_oyster_18S_V4}
