% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.db.byid.R
\name{search.db.byid}
\alias{search.db.byid}
\title{A search.db.byid Function}
\usage{
search.db.byid(query, cas = NULL)
}
\arguments{
\item{query}{character, the name of gRNA-target-site, such as "probe_022593".}
}
\description{
This function allows you to search the target-site oligonucletide sequence of gRNA in the gRNA-targe-site database based on the input ID of the gRNA-target-site.
}
\examples{
search.db.byid(query="probe_022593", cas="Cas9")

}
\keyword{search.db.byid}
