% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.db.byname.R
\name{search.db.byname}
\alias{search.db.byname}
\title{A search.db.byname Function}
\usage{
search.db.byname(query, cas = NULL)
}
\arguments{
\item{query}{character, the name of host species or taxonomic group (from species to kingdom) that you expect the gRNA to target.}
}
\description{
This function allows you to search the target-site oligonucletide sequence of gRNA in the gRNA-targe-site database based on the input name of the host species.
}
\examples{
search.db.byname(query="Homo sapiens", cas="Cas9")
search.db.byname(query="Salmon", cas="Cas9")
search.db.byname(query="Mollusca", cas="Cas9")
search.db.byname(query="Crassostrea", cas="Cas9")
Please be noted that the name of host species or taxonomic group needs to be a legistimal names appeared in SILVA SSU 18S database.
}
\keyword{search.db.byname}
