% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.silva_v119_18S_99rep_V4.R
\docType{data}
\name{silva_v119_18S_99rep_V4}
\alias{silva_v119_18S_99rep_V4}
\title{silva_v119_18S_99rep_V4:
Reference database of the 18S rRNA gene at V4 region amplified by primers TAReuk454FWD1 and TAReukREV3 (Stoeck et al., 2010), which uses for the identify the gRNA-target-sequence oligos.}
\format{
A DNAString fasta input
}
\usage{
data(silva_v119_18S_99rep_V4)
}
\description{
silva_v119_18S_99rep_V4:
Reference database of the 18S rRNA gene at V4 region amplified by primers TAReuk454FWD1 and TAReukREV3 (Stoeck et al., 2010), which uses for the identify the gRNA-target-sequence oligos.
}
\examples{
data(silva_v119_18S_99rep_V4)
}
\references{
Stoeck T, Bass D, Nebel M, Christen R, Jones MD, Breiner HW, Richards TA. (2010)
Multiple marker parallel tag environmental DNA sequencing reveals a highly complex eukaryotic community in marine anoxic water
Mol. Ecol. 19, 21–31.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/20331767}{PubMed})
}
\keyword{silva_v119_18S_99rep_V4}
