% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.db.byname.R
\name{search.db.byname}
\alias{search.db.byname}
\title{A search.db.byname Function}
\usage{
search.db.byname(query, cas = NULL)
}
\arguments{
\item{query}{character, the name of host species or taxonomic group (from species to kingdom) that you expect the gRNA to target.}
}
\description{
This function allows you to search the DNA oligo for guding RNA for CRISPR-cas9 and/or CRISPR-Cas12a system with aim to specifically cut 18S rRNA genes sequences of hosts but not protists and fungus. The database is build based on all available taxa in SILVA 18S SSU database version 119 (July 2014)
}
\examples{
search.db.byname(query="Homo sapiens")
search.db.byname(query="Salmon")
search.db.byname(query="Mollusca")
search.db.byname(query="Crassostrea")
Please be noted that the name of host species or taxonomic group needs to be a legistimal names appeared in SILVA SSU 18S database.
}
\keyword{search.db.byname}
